requirements.txt
python-telegram-bot==21.6
openai>=1.40.0
SQLAlchemy>=2.0.30
pydantic>=2.7.0
python-dotenv>=1.0.1
tzdata>=2024.1 